requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.0.0
Pillow>=8.0.0
opencv-python>=4.5.0
numpy>=1.19.0
